requirements.txt
dash
pandas
plotly
google-api-python-client
python-dotenv
isodate